refactor(economy): phase 3 slice 30 validator split and runner merge (#3939)
Split the monolithic validator scenario module into behavior-family files,
relocate price-discovery scenarios under the treasury cluster, and merge
purchased-tile runners into the consolidated context test file.

Refs #3939
