+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -37,8 +37,6 @@ export 'src/tile_extraction_contribution_scenarios.dart';
 export 'src/tile_extraction_pipeline_scenarios.dart';
 export 'src/trade_cargo_capacity_scenarios.dart';
 export 'src/treasury_scenarios/treasury_scenarios.dart';
-export 'src/price_discovery_expectations.dart';
-export 'src/price_discovery_scenarios.dart';
 export 'src/projected_cost_engine_scenarios.dart';
 export 'src/purchased_tile_fixture_support.dart';
 export 'src/purchased_tile_scenarios/purchased_tile_expectations.dart';
