+++ packages/colonizethis_economy_test_support/lib/src/treasury_scenarios/price_discovery_scenarios.dart
