+++ packages/colonizethis_economy_test_support/lib/src/treasury_scenarios/treasury_scenarios.dart
@@ -8,5 +8,7 @@ export 'treasury_player_context_scenarios.dart';
 export 'treasury_pure_helper_scenarios.dart';
 export 'treasury_sellable_quantity_scenarios.dart';
 export 'treasury_test_support.dart';
+export 'price_discovery_expectations.dart';
+export 'price_discovery_scenarios.dart';
 export 'world_market_context_base_expectations.dart';
 export 'world_market_context_base_scenarios.dart';
