+++ packages/colonizethis_economy_test_support/lib/src/validator_scenarios/validator_scenario.dart
@@ -1,0 +1,45 @@
+// TradeOrderValidator scenario row type and runner (Refs #3836, #3939 phase 3 slice 30).
+
+import 'package:colonizethis_economy/colonizethis_economy.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'validator_expectations.dart';
+
+/// One row for validator cap / rules / treasury scenarios.
+class TradeOrderValidatorScenario {
+  const TradeOrderValidatorScenario({
+    required this.label,
+    required this.context,
+    required this.proposedOrders,
+    required this.verify,
+    this.refs,
+  });
+
+  TradeOrderValidatorScenario.expect({
+    required String label,
+    required TradeOrderValidationContext context,
+    required List<TradeOrder> proposedOrders,
+    required ValidatorExpectation expect,
+    String? refs,
+  }) : this(
+          label: label,
+          context: context,
+          proposedOrders: proposedOrders,
+          verify: (results) => assertValidatorExpectation(results, expect),
+          refs: refs,
+        );
+
+  final TradeOrderValidationContext context;
+  final List<TradeOrder> proposedOrders;
+  final String label;
+  final void Function(List<OrderValidationResult> results) verify;
+  final String? refs;
+}
+
+void runTradeOrderValidatorScenario(TradeOrderValidatorScenario scenario) {
+  final results = TradeOrderValidator.validate(
+    context: scenario.context,
+    proposedOrders: scenario.proposedOrders,
+  );
+  scenario.verify(results);
+}
