+++ packages/colonizethis_economy_test_support/lib/src/validator_scenarios/validator_scenarios.dart
@@ -1,5 +1,9 @@
-// Barrel for consolidated TradeOrderValidator scenario modules (Refs #3939 phase 3 slice 7).
+// Barrel for consolidated TradeOrderValidator scenario modules (Refs #3939 phase 3 slice 7+).
 
-export 'trade_order_validator_scenarios.dart';
 export 'trade_order_validator_test_support.dart';
+export 'validator_cap_scenarios.dart';
+export 'validator_context_scenarios.dart';
 export 'validator_expectations.dart';
+export 'validator_rules_scenarios.dart';
+export 'validator_scenario.dart';
+export 'validator_treasury_scenarios.dart';
